WIP(lexvalidator): rename
