+++ src/plccng/lexparse/__init__.py
@@ -1,10 +1,10 @@
+from .lexvalidate.PatternValidator import InvalidPattern
+from .lexvalidate.NameValidator import InvalidName
+from .lexvalidate.UniqueNameValidator import DuplicateName
 from .LexicalSpec import LexicalSpec
 from .LexicalRule import LexicalRule
-from .validations import (
-    DuplicateRuleName,
-    InvalidRuleName,
-    InvalidRulePattern,
-    InvalidRule,
+from .lexvalidate.UnrecognizedLineValidator import (
+    UnrecognizedLine,
 )
 from .parse_lexical_spec import from_string
 from .parse_lexical_spec import from_lines
