--- src/plccng/lexparse/validations/DuplicateRuleName.py
@@ -1,9 +1,0 @@
-from dataclasses import dataclass
-from plccng.ValidationError import ValidationError
-
-
-@dataclass
-class DuplicateRuleName(ValidationError):
-    def __init__(self, rule):
-        self.line = rule.line
-        self.message = f"Duplicate rule name found '{rule.name}' on line: {rule.line.number}"
