--- src/plccng/lexparse/validations/InvalidRule.py
@@ -1,9 +1,0 @@
-from dataclasses import dataclass
-from plccng.ValidationError import ValidationError
-
-
-@dataclass
-class InvalidRule(ValidationError):
-    def __init__(self, line):
-        self.line = line
-        self.message = f"Invalid rule format found on line: {line.number}"
